moving contents of autoignition back into simulation file
